irsa/jwt_validation/requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
idna==3.7
install==1.3.5
pycparser==2.22
PyJWT==2.8.0
requests==2.31.0
urllib3==2.2.1
